requirements.txt
Flask
torch
torchvision
Pillow
Flask-Cors